+++ include/mse/systems/windows/layers/gui/items/button.h
@@ -19,19 +19,28 @@ namespace mse
 		public:
 			// general initialization
 			Button();
-			Button(Layer* layer, const std::string& text, const glm::uvec4& area, const glm::uvec4& bgColor, const glm::uvec4& color);
-			void Init(Layer* layer,  const std::string& text, const glm::uvec4& area, const glm::uvec4& bgColor, const glm::uvec4& color);
+			Button(Layer* layer, const std::u32string& text, const glm::uvec4& area, const glm::uvec4& bgColor, const glm::uvec4& color);
+			void Init(Layer* layer,  const std::u32string& text, const glm::uvec4& area, const glm::uvec4& bgColor, const glm::uvec4& color);
 			virtual ~Button();
 			
 			// general GUIItem interface
 			virtual void Display();
+			virtual bool HandleEvent(EventTypes eventType, SDL_Event* event) override;
+
+			inline void Enable() { isEnabled = true; }
+			inline void Disable() { isEnabled = false; }
+			inline void ToggleEnabled() { isEnabled = !isEnabled; }
 			
 			// unique Button interface
 			ButtonStates state = ButtonStates::Released;
 		protected:
-			std::string m_text = "";
+			std::u32string m_text = U"";
 			glm::uvec4 m_backgroundColor = {0, 0, 0, 0};
 			glm::uvec4 m_textColor = {223, 223, 223, 255};
+			
+			bool isHover = false;
+			bool isPushed = false;
+			bool isEnabled = true;
 		};
 	}
 }