+++ tests/logger-testbed.cpp
@@ -3,70 +3,269 @@
 class SimpleUILayer : public mse::Layer
 {
 public:
+	enum class MovementDirection
+	{
+		None,
+		Left,
+		Right,
+		Down,
+		Up
+	};
+	
 	SimpleUILayer()
 	: mse::Layer()
 	{
 		MSE_LOG("Constructed a simpleUI layer");
+		
+		gameBoard.resize(10 * 20, 0);
+		gameBoard[12] = 1;
 	}
 	
 	virtual void OnInit() override
 	{
-		mse::gui::Canvas* canvas = (mse::gui::Canvas*)(AddElement(new mse::gui::Canvas(this, {20, 10, 290, 220}, {32, 64, 48, 255})));
-		canvas->SetDrawColor({228, 228, 228, 255});
-
-		mse::Resource* bmpFont = mse::ResourceManager::UseResource(mse::ResourceType::FontBitmap, "./data/fonts/my8bit2.png", this->GetWindow());
-		mse::Texture* canvasTexture = (mse::Texture*)(canvas->GetTexture()->data);
-		mse::Renderer::SurfaceDrawText(
-			canvasTexture, 		// what element to draw on
-			{30, 30, 200, 80}, 	// where to
-			2, 					// pixel size
-			U"Привет, мир!", 	// text content
-			bmpFont, 			// font
-			{255, 255, 0, 255}, // color
-			0); 				// interval between rows
-		mse::Renderer::SurfaceDrawText(
-			canvasTexture, 		// what element to draw on
-			{10, 60, 250, 120}, // where to
-			1, 					// pixel size
-			U"У вас дислексия? Вы не выговари-ваете букву \"р\"? Новое средство!\nИммунооцелистонатриихлорогидро-\nкарбодон!\nСпрашивайте в аптеках города.", 	// text content
-			bmpFont, 			// font
-			{196, 196, 255, 255}, // color
-			2); 				// interval between rows
-		canvasTexture->Update();
+		bmpFont = mse::ResourceManager::UseResource(mse::ResourceType::FontBitmap, "./data/fonts/my8bit3.bmp", this->GetWindow());
 		
-		mse::gui::Button* redPencil = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, "Red Pencil", {2, 10, 15, 10}, {196, 64, 64, 255}, {196, 196, 32, 255})));
-		redPencil->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&, canvas](SDL_Event* event){
-			canvas->SetDrawColor({228, 64, 64, 255});
+		// Game GUI
+		gameCanvas = (mse::gui::Canvas*)(AddElement(new mse::gui::Canvas(this, {10, 10, gameCanvasGrid.x * gameCanvasGrid.z, gameCanvasGrid.y * gameCanvasGrid.w}, {32, 64, 48, 255})));
+		elementInFocus = GetElementId(gameCanvas);
+		MSE_LOG(elementInFocus);
+		
+		gameCanvas->callbacks[mse::EventTypes::KeyDown] = [&](SDL_Event* event){
+			MSE_LOG("Callback!");
+			switch (event->key.keysym.sym)
+			{
+				case SDLK_LEFT:
+				{
+					direction = MovementDirection::Left;
+					break;
+				}
+				case SDLK_RIGHT:
+				{
+					direction = MovementDirection::Right;
+					break;
+				}
+				case SDLK_DOWN:
+				{
+					direction = MovementDirection::Down;
+					break;
+				}
+				case SDLK_UP:
+				{
+					direction = MovementDirection::Up;
+					break;
+				}
+			}
+			MSE_LOG("Player: ", player.x, " : ", player.y);
 		};
-		mse::gui::Button* greenPencil = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, "Green Pencil", {2, 20, 15, 10}, {64, 196, 64, 255}, {196, 196, 32, 255})));
-		greenPencil->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&, canvas](SDL_Event* event){
-			canvas->SetDrawColor({64, 228, 64, 255});
+
+		mse::gui::Button* playBtn = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, U"Играть", {100, 10, 80, 10}, {196, 196, 196, 255}, {32, 32, 32, 255})));
+		playBtn->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&](SDL_Event* event){
+			mode = 1;
 		};
-		mse::gui::Button* bluePencil = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, "Blue Pencil", {2, 30, 15, 10}, {64, 64, 196, 255}, {196, 196, 32, 255})));
-		bluePencil->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&, canvas](SDL_Event* event){
-			canvas->SetDrawColor({64, 64, 228, 255});
+		mse::gui::Button* settingsBtn = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, U"Настройки", {100, 20, 80, 10}, {196, 196, 196, 255}, {32, 32, 32, 255})));
+		settingsBtn->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&](SDL_Event* event){
+			mode = 2;
 		};
-		mse::gui::Button* yellowPencil = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, "Yellow Pencil", {2, 40, 15, 10}, {196, 196, 64, 255}, {196, 196, 32, 255})));
-		yellowPencil->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&, canvas](SDL_Event* event){
-			canvas->SetDrawColor({228, 228, 64, 255});
+		mse::gui::Button* scoresBtn = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, U"Рекорды", {100, 30, 80, 10}, {196, 196, 196, 255}, {32, 32, 32, 255})));
+		scoresBtn->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&](SDL_Event* event){
+			mode = 3;
 		};
-		mse::gui::Button* pinkPencil = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, "Pink Pencil", {2, 50, 15, 10}, {196, 64, 196, 255}, {196, 196, 32, 255})));
-		pinkPencil->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&, canvas](SDL_Event* event){
-			canvas->SetDrawColor({228, 64, 228, 255});
+		mse::gui::Button* aboutBtn = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, U"Титры", {100, 40, 80, 10}, {196, 196, 196, 255}, {32, 32, 32, 255})));
+		aboutBtn->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&](SDL_Event* event){
+			mode = 4;
 		};
-		mse::gui::Button* cyanPencil = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, "Cyan Pencil", {2, 60, 15, 10}, {64, 196, 196, 255}, {196, 196, 32, 255})));
-		cyanPencil->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&, canvas](SDL_Event* event){
-			canvas->SetDrawColor({64, 228, 228, 255});
+		mse::gui::Button* exitBtn = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, U"Выйти", {100, 50, 80, 10}, {196, 196, 196, 255}, {32, 32, 32, 255})));
+		exitBtn->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&](SDL_Event* event){
+			mode = 5;
 		};
-		mse::gui::Button* whitePencil = (mse::gui::Button*)(AddElement(new mse::gui::Button(this, "White Pencil", {2, 70, 15, 10}, {228, 228, 228, 255}, {32, 32, 32, 255})));
-		whitePencil->callbacks[mse::EventTypes::GUIItemMouseButtonUp] = [&, canvas](SDL_Event* event){
-			canvas->SetDrawColor({228, 228, 228, 255});
-		};
+		
+		scoreCanvas = (mse::gui::Canvas*)(AddElement(new mse::gui::Canvas(this, {100, 70, 80, 30}, {32, 64, 48, 255})));
+		
+		itemCanvas = (mse::gui::Canvas*)(AddElement(new mse::gui::Canvas(this, {100, 110, 80, 60}, {32, 64, 48, 255})));
+		
+		mse::Application::GetApplication()->SetFPS(MSE_FPS4);
+	}
+	
+	virtual void OnUpdate() override
+	{
+		switch (mode)
+		{
+			// idle
+			case 0:
+			{
+				break;
+			}
+			// game
+			case 1:
+			{
+				if (!gameOver)
+				{
+					gameCanvas->Clear();
+					scoreCanvas->Clear();
+					itemCanvas->Clear();
+					
+					// correct player position
+					player.x = (player.x < 0) ? 0 : player.x;
+					player.x = (player.x > gameCanvasGrid.z - 1) ? gameCanvasGrid.z - 1 : player.x;
+					player.y = (player.y > gameCanvasGrid.w - 1) ? gameCanvasGrid.w - 1 : player.y;
+					player.y = (player.y < 0) ? 0 : player.y;
+					
+					// update player
+					MovePlayer(direction);
+					
+					// update tail
+					for (int i = tail.size() - 1; i > 0; --i)
+					{
+						tail[i] = tail[i - 1];
+						
+						// check if the player collides with his tail 
+						if ((tail[i].x == player.x) && (tail[i].y == player.y))
+						{
+							gameOver = true;
+						}
+					}
+					
+					if ((player.x == apple.x) && (player.y == apple.y))
+					{
+						// increase score
+						player.z++;
+						
+						// increase tail
+						tail.push_back({player.x, player.y});
+						
+						apple.x = rand() % gameCanvasGrid.z;
+						apple.y = rand() % gameCanvasGrid.w;
+					}
+					
+					// draw tail
+					for (int i = tail.size() - 1; i > 0; --i)
+					{
+						gameCanvas->DrawRect(
+							tail[i].x*gameCanvasGrid.x, 
+							tail[i].y*gameCanvasGrid.y, 
+							(tail[i].x + 1)*gameCanvasGrid.x, 
+							(tail[i].y + 1)*gameCanvasGrid.y, 
+							{255, 255, 0, 255}
+							);
+					}
+					
+					// draw player
+					gameCanvas->DrawRect(
+						player.x*gameCanvasGrid.x, 
+						player.y*gameCanvasGrid.y, 
+						(player.x + 1)*gameCanvasGrid.x, 
+						(player.y + 1)*gameCanvasGrid.y, 
+						{255, 0, 255, 255}
+						);
+					
+					// draw apple
+					gameCanvas->DrawRect(
+						apple.x*gameCanvasGrid.x, 
+						apple.y*gameCanvasGrid.y, 
+						(apple.x + 1)*gameCanvasGrid.x, 
+						(apple.y + 1)*gameCanvasGrid.y, 
+						{255, 0, 0, 255}
+						);
+					
+					// draw scores
+					std::stringstream score;
+					score << player.z;
+					mse::Renderer::SurfaceDrawText(
+						(mse::Texture*)(scoreCanvas->GetTexture()->data),
+						{5, 5, 70, 20},
+						1,
+//						U"Hello!",
+						U"Счёт: " + utf8::utf8to32(score.str()),
+						bmpFont,
+						{255, 255, 0, 255},
+						0
+						);
+					score.str("");
+					score << player.w;
+					mse::Renderer::SurfaceDrawText(
+						(mse::Texture*)(scoreCanvas->GetTexture()->data),
+						{5, 15, 70, 20},
+						1,
+//						U"Hello!",
+						U"Жизни: " + utf8::utf8to32(score.str()),
+						bmpFont,
+						{255, 255, 0, 255},
+						0
+						);
+					((mse::Texture*)(scoreCanvas->GetTexture()->data))->Update();
+				}
+				
+				break;
+			}
+			// settings
+			case 2:
+			{
+				break;
+			}
+			// scores
+			case 3:
+			{
+				break;
+			}
+			// about
+			case 4:
+			{
+				break;
+			}
+			// exit
+			case 5:
+			{
+				MSE_LOG("Commanding to stop the application");
+				mse::Application::GetApplication()->Stop();
+				break;
+			}
+		}
 	}
 	
 	~SimpleUILayer()
 	{
 		MSE_LOG("Destroyed a simpleUI layer");
+	}
+	
+	int mode = 0; // 0 - nothing, 1 - game, 2 - settings, 3 - scores, 4 - about, 5 - exit
+	std::vector<int> gameBoard = {};
+	mse::gui::Canvas* gameCanvas = nullptr;
+	mse::gui::Canvas* scoreCanvas = nullptr;
+	mse::gui::Canvas* itemCanvas = nullptr;
+	glm::ivec4 gameCanvasGrid = {8, 8, 10, 20};
+	glm::ivec4 player = {5, 10, 1, 3}; // x, y, length, lives
+	MovementDirection direction = MovementDirection::None; // 1 - left, 2 - right, 3 - down, 4 - up;
+	std::vector<glm::ivec2> tail = {{-1, -1}};
+	glm::ivec2 apple = {0, 0};
+	bool gameOver = false;
+	mse::Resource* bmpFont = nullptr;
+	
+	void MovePlayer(MovementDirection dir)
+	{
+		tail[0] = {player.x, player.y};
+		switch (dir)
+		{
+			case MovementDirection::Left:
+			{
+				--player.x;
+				break;
+			}
+			case MovementDirection::Right:
+			{
+				++player.x;
+				break;
+			}
+			case MovementDirection::Up:
+			{
+				--player.y;
+				break;
+			}
+			case MovementDirection::Down:
+			{
+				++player.y;
+				break;
+			}
+		}
 	}
 };
 
@@ -79,7 +278,7 @@ public:
 		MSE_CORE_ERROR("Joke");
 		
 		MSE_LOG("Commanding to open a window");
-		m_window = mse::WindowManager::CreateWindow(u8"Test window", 50, 50, 320, 240);
+		m_window = mse::WindowManager::CreateWindow(u8"Test window", 50, 50, 200, 200);
 
 		mse::Renderer::SetActiveWindow(m_window);
 		m_window->GetLayerManager()->Attach(new SimpleUILayer());